AC-22 #comment reorganized components
